+++ package.json
@@ -11,6 +11,7 @@
     "nanoid": "^1.0.2",
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.4",
     "typeface-roboto": "0.0.54"
   },
