+++ src/components/WelcomePage.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const WelcomePage = () => (
+  <h1>Hello User!</h1>
+);
+
+export default WelcomePage;
